% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkQuartet.R
\name{checkQuartet}
\alias{checkQuartet}
\title{checkQuartet}
\usage{
checkQuartet(tips = c(1, 2, 3, 4), readSeq)
}
\arguments{
\item{tips}{Defaults to c(1,2,3,4)}

\item{readSeq}{a list of vectors of same length}
}
\value{
bestConfiguration, flag, caseLength
}
\description{
This function finds the best quartet given three tip indices and a phyDat list of sequences
}

