% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherryC.R
\name{cherryC}
\alias{cherryC}
\title{cherryC}
\usage{
cherryC(tipC, readSeq)
}
\arguments{
\item{tipC}{numeric index of the tip}

\item{readSeq}{a list of vectors of same length}
}
\value{
bestConfiguration, flag, caseLength
}
\description{
This function determines if tipC is in a cherry and che other tips in the cherry
}

