% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT-package.R
\docType{package}
\name{CWT-package}
\alias{CWT}
\alias{CWT-package}
\title{Continuos Wavelet Transformation for Spectroscopy}
\description{
Fast application of Continous Wavelet Transformation on time
series with special attention to spectroscopy. It is written using
data.table and C++ language and in some functions it is possible to
use parallel processing to speed-up the computation over samples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Antguz/CWT}
  \item Report bugs at \url{https://github.com/Antguz/CWT/issues}
}

}
\author{
\strong{Maintainer}: J. Antonio Guzmán Q. \email{antguz06@gmail.com} (\href{https://orcid.org/0000-0002-0721-148X}{ORCID}) [copyright holder]

}
