% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDist.R
\name{genDist}
\alias{genDist}
\title{genDist}
\usage{
genDist(tip1, tip2)
}
\arguments{
\item{tip1}{a vector}

\item{tip2}{another vector}
}
\value{
distance
}
\description{
This function calculates the Jukes/Cantor distance between two vectors
}

