% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialiseC.R
\name{initialiseC}
\alias{initialiseC}
\title{initialiseC}
\usage{
initialiseC(tipC, readSeq)
}
\arguments{
\item{tipC}{numeric index of a tip}

\item{readSeq}{a list of vectors of same length}
}
\value{
bestConfiguration, tipM, tipF, tipT, bestConfiguration, flag, caseLength
}
\description{
This internal function is used to initialise the cherryC function
}

