% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcwt.R
\name{parcwt}
\alias{parcwt}
\title{parcwt}
\usage{
parcwt(sequences)
}
\arguments{
\item{sequences}{a list of vectors of same length}
}
\value{
cherryNumber
}
\description{
This function finds the number of the cherries given a phyDat list of sequences
}

